backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<4
djangorestframework-simplejwt>=5.3,<6
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
Pillow>=10,<12
django-cors-headers>=4.4,<5
gunicorn>=22,<23
